refactor: 숫자 문자열을 Numbers 클래스로 분리
